requirements.txt
python-telegram-bot
psutil
opencv-python
Pillow